IS-2994: Vis sykepengesoknader sykmeldt har vaert i arbeid
